+++ src/main/java/net/mcreator/orderofobsidian/procedures/CoalGeneratorFuelBar1Procedure.java
@@ -7,14 +7,14 @@ import net.minecraft.core.BlockPos;
 public class CoalGeneratorFuelBar1Procedure {
 	public static boolean execute(LevelAccessor world, double x, double y, double z) {
 		double craftingTime = 0;
-		craftingTime = (double) ((new Object() {
+		craftingTime = (new Object() {
 			public double getValue(LevelAccessor world, BlockPos pos, String tag) {
 				BlockEntity blockEntity = world.getBlockEntity(pos);
 				if (blockEntity != null)
 					return blockEntity.getTileData().getDouble(tag);
 				return -1;
 			}
-		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "burningTime")) / 10);
+		}.getValue(world, new BlockPos((int) x, (int) y, (int) z), "burningTime")) / 10;
 		if (new Object() {
 			public double getValue(LevelAccessor world, BlockPos pos, String tag) {
 				BlockEntity blockEntity = world.getBlockEntity(pos);
